pagination

+++ controllers/admin.js
@@ -1,3 +1,4 @@
+const fileHelper = require("../util/file");
 const Product = require("../models/product");
 const { validationResult } = require("express-validator/check");
 
@@ -63,20 +64,6 @@ exports.postAddProduct = (req, res, next) => {
       res.redirect("/admin/products");
     })
     .catch((err) => {
-      // return res.status(422).render("admin/edit-product", {
-      //   pageTitle: "Add Product",
-      //   path: "/admin/add-product",
-      //   editing: false,
-      //   hasError: true,
-      //   product: {
-      //     title: title,
-      //     price: price,
-      //     imageUrl: imageUrl,
-      //     description: description,
-      //   },
-      //   errorMessage: "Database operation failed, please try again.",
-      //   validationErrors: [],
-      // });
       const errors = new Error(err);
       errors.httpStatus = 500;
       return next(errors);
@@ -143,6 +130,7 @@ exports.postEditProduct = (req, res, next) => {
       product.price = updatedPrice;
       product.description = updatedDesc;
       if (image) {
+        fileHelper.deleteFile(product.imageUrl);
         product.imageUrl = image.path;
       }
       return product.save().then((result) => {
@@ -178,14 +166,21 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  Product.deleteOne({ _id: prodId, userId: req.user._id })
+  Product.findById(prodId)
+    .then((product) => {
+      if (!product) {
+        return next(new Error("Product not found."));
+      }
+      fileHelper.deleteFile(product.imageUrl);
+      return Product.deleteOne({ _id: prodId, userId: req.user._id });
+    })
     .then(() => {
       console.log("DESTROYED PRODUCT");
       res.redirect("/admin/products");
     })
     .catch((err) => {
-      const errors = new Error(err);
-      errors.httpStatus = 500;
-      return next(errors);
+      const error = new Error(err);
+      error.httpStatusCode = 500;
+      return next(error);
     });
 };
